DjangoRestFramework/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.7
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
filelock==3.13.1
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
pipenv==2023.11.15
platformdirs==4.0.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
virtualenv==20.24.7
